# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WrzDJ API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/events/{code}/requests": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Submit Request",
        "operationId": "submit_request_api_events__code__requests_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event Requests",
        "operationId": "get_event_requests_api_events__code__requests_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "chronological",
                "priority"
              ],
              "type": "string",
              "default": "chronological",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestOut"
                  },
                  "title": "Response Get Event Requests Api Events  Code  Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/events/{code}/display": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Kiosk Display",
        "description": "Get public kiosk display data for an event.",
        "operationId": "get_kiosk_display_api_public_events__code__display_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KioskDisplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/events/{code}/requests": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Requests",
        "description": "Get publicly visible requests for an event (NEW and ACCEPTED only).",
        "operationId": "get_public_requests_api_public_events__code__requests_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/events/{code}/has-requested": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Check Has Requested",
        "description": "Check if the current client has submitted any requests for this event.",
        "operationId": "check_has_requested_api_public_events__code__has_requested_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HasRequestedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/events/{code}/my-requests": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get My Requests",
        "description": "Get all requests submitted by the current client for this event.",
        "operationId": "get_my_requests_api_public_events__code__my_requests_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRequestsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuestNowPlaying": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "album_art_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Art Url"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "title",
          "artist",
          "album_art_url",
          "source"
        ],
        "title": "GuestNowPlaying"
      },
...
      "GuestRequestListResponse": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/PublicEventInfo"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/GuestRequestInfo"
            },
            "type": "array",
            "title": "Requests"
          },
          "now_playing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuestNowPlaying"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "event",
          "requests"
        ],
        "title": "GuestRequestListResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HasRequestedResponse": {
        "properties": {
          "has_requested": {
            "type": "boolean",
            "title": "Has Requested"
          }
        },
        "type": "object",
        "required": [
          "has_requested"
        ],
        "title": "HasRequestedResponse"
      },
...
      "KioskDisplayResponse": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/PublicEventInfo"
          },
          "qr_join_url": {
            "type": "string",
            "title": "Qr Join Url"
          },
          "accepted_queue": {
            "items": {
              "$ref": "#/components/schemas/PublicRequestInfo"
            },
            "type": "array",
            "title": "Accepted Queue"
          },
          "now_playing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicRequestInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "now_playing_hidden": {
            "type": "boolean",
            "title": "Now Playing Hidden"
          },
          "requests_open": {
            "type": "boolean",
            "title": "Requests Open",
            "default": true
          },
          "kiosk_display_only": {
            "type": "boolean",
            "title": "Kiosk Display Only",
            "default": false
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "banner_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Url"
          },
          "banner_kiosk_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Kiosk Url"
          },
          "banner_colors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Colors"
          }
        },
        "type": "object",
        "required": [
          "event",
          "qr_join_url",
          "accepted_queue",
          "now_playing",
          "now_playing_hidden",
          "updated_at"
        ],
        "title": "KioskDisplayResponse"
      },
...
      "MyRequestsResponse": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/MyRequestInfo"
            },
            "type": "array",
            "title": "Requests"
          }
        },
        "type": "object",
        "required": [
          "requests"
        ],
        "title": "MyRequestsResponse"
      },
...
      "PublicEventInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "PublicEventInfo"
      },
      "PublicRequestInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "vote_count": {
            "type": "integer",
            "title": "Vote Count",
            "default": 0
          },
          "bpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "musical_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Musical Key"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "requester_verified": {
            "type": "boolean",
            "title": "Requester Verified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "artwork_url"
        ],
        "title": "PublicRequestInfo"
      },
...
      "RequestCreate": {
        "properties": {
          "artist": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Artist"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "source": {
            "$ref": "#/components/schemas/RequestSource",
            "default": "manual"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "raw_search_query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Search Query"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "musical_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Musical Key"
          }
        },
        "type": "object",
        "required": [
          "artist",
          "title"
        ],
        "title": "RequestCreate"
      },
      "RequestOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "song_title": {
            "type": "string",
            "title": "Song Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "is_duplicate": {
            "type": "boolean",
            "title": "Is Duplicate",
            "default": false
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "musical_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Musical Key"
          },
          "raw_search_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Search Query"
          },
          "sync_results_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Results Json"
          },
          "vote_count": {
            "type": "integer",
            "title": "Vote Count",
            "default": 0
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "song_title",
          "artist",
          "source",
          "source_url",
          "artwork_url",
          "note",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "RequestOut"
      },
      "RequestSource": {
        "type": "string",
        "enum": [
          "manual",
          "musicbrainz",
          "spotify",
          "share_link",
          "tidal",
          "beatport"
        ],
        "title": "RequestSource"
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "new",
          "accepted",
          "playing",
          "played",
          "rejected"
        ],
        "title": "RequestStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}